Switching delta_time type to float

+++ IgniteEngine/IgniteEngine/DefaultConf.cpp
@@ -16,7 +16,7 @@ Queue* DefaultConf::graphics_queue;
 Queue* DefaultConf::present_queue;
 Queue* DefaultConf::compute_queue;
 Texture* DefaultConf::white_texture;
-uint64_t DefaultConf::delta_time; // ms
+float DefaultConf::delta_time; // ms
 SDL_Event* DefaultConf::event;
 
 VertexBuffer* DefaultConf::coord_buffer;
